bronze/requirements.txt
google-cloud-storage
gcsfs
great_expectations
pandas
google-cloud-bigquery
apache-airflow
fastapi
uvicorn